update climat component
