+++ src/components/ModalCountryDetails/Places/Slider/Slider.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import "react-responsive-carousel/lib/styles/carousel.min.css";
 import { Carousel } from "react-responsive-carousel";
 import "./slider.css";
@@ -10,20 +10,10 @@ interface MustSeePlace {
 }
 
 interface CountryProp {
-	country: {
-		translations: {
-			fra: {
-				common: string;
-			};
-		};
-	};
-}
-interface Country {
-	country: string;
-	must_see_places: MustSeePlace[];
+	translations: Record<string, { common: string }>;
 }
 
-function Slider({ country }: CountryProp) {
+function Slider({ country }: { country: CountryProp }) {
 	const [myPlaces, setMyPlaces] = useState<MustSeePlace[]>([]);
 
 	useEffect(() => {
