+++ src/components/ModalCountryDetails/Places/Slider/slider.css
@@ -1,20 +1,3 @@
-ul li {
-	list-style-type: none;
-	border-radius: 15px;
-	border: solid rgb(66, 140, 149);
-	box-shadow: 20px 20px 20px 0 rgb(66, 140, 149, 0.6);
-}
-
-ul img {
-	border-radius: 5px;
-}
-
-ul li:hover {
-	transform: scale(1.05);
-	transition: all 0.3s ease;
-	background: linear-gradient(to bottom, rgb(66, 140, 149), white);
-	border-radius: 15px;
-}
 .carousel {
 	width: 60%;
 	margin: auto;
