+++ src/pages/InteractiveMap/ContinentModal/ContinentModal.css
@@ -21,9 +21,10 @@
 }
 
 .continent-modal-content {
+	display: none;
 	background-color: #ffffff7a;
 	border-radius: 10px;
-	padding: 20px 30px;
+	padding: 10px 20px;
 	width: 90%;
 	max-width: 400px;
 	box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
@@ -53,11 +54,11 @@
 }
 
 .close-modal-continent:hover {
-	color: #e57373;
+	color: #9b7316;
 }
 
 .continent-modal-content h3 {
-	font-size: 1.8rem;
+	font-size: 1.3rem;
 	margin-bottom: 20px;
 	text-align: center;
 	color: #333;
@@ -65,30 +66,29 @@
 
 .continent-list {
 	display: grid;
-	grid-template-columns: 1fr 1fr;
+	grid-template-columns: repeat(2, 1fr);
 	gap: 15px;
-	margin-top: 20px;
 }
 
 .continent-region {
 	padding: 15px;
-	font-size: 1.2rem;
+	font-size: 1rem;
 	background-color: #4caf50;
 	color: #fff;
 	border: none;
 	border-radius: 8px;
 	cursor: pointer;
 	text-align: center;
-	transition: background-color 0.3s ease, transform 0.3s ease;
+	transition: all 0.3s ease;
 }
 
 .continent-region:hover {
 	background-color: #388e3c;
-	transform: translateY(-3px);
+	transform: scale(1.02);
 }
 
 @media (max-width: 768px) {
-	.continent-modal-overlay {
+	.continent-modal {
 		background: rgba(0, 0, 0, 0.5);
 	}
 	.continent-modal-content {
